+++ packages/admin-ui/src/lib/catalog/src/components/generate-product-variants/generate-product-variants.component.ts
@@ -1,5 +1,5 @@
 import { Component, ElementRef, EventEmitter, OnInit, Output, QueryList, ViewChildren } from '@angular/core';
-import { AbstractControl, FormArray, FormBuilder, FormControl, FormGroup } from '@angular/forms';
+import { AbstractControl, FormBuilder, FormControl, FormGroup } from '@angular/forms';
 import {
     CurrencyCode,
     DataService,
@@ -70,6 +70,7 @@ export class GenerateProductVariantsComponent implements OnInit {
                 tap(items => {
                     if (items.length) {
                         this.selectedStockLocationId = items[0].id;
+                        this.onFormChange();
                     }
                 }),
             );