+++ packages/dev-server/example-plugins/ui-extensions-library/ui/routes.ts
@@ -1,4 +1,4 @@
-import { registerRouteComponent } from '@vendure/admin-ui/core';
+import { CanDeactivateDetailGuard, registerRouteComponent } from '@vendure/admin-ui/core';
 import { registerReactRouteComponent } from '@vendure/admin-ui/react';
 
 import { AngularUiComponent } from './angular-components/angular-ui/angular-ui.component';
@@ -14,5 +14,8 @@ export default [
         path: 'angular-ui',
         component: AngularUiComponent,
         title: 'Angular UI',
+        routeConfig: {
+            canDeactivate: [CanDeactivateDetailGuard],
+        },
     }),
 ];